Remove "r_mode 2" for ioquake3 and quake3e
The command argument `r_mode -2` for both ioquake3 and quake3e is redundant given that it's the default mode when the config file is created (in `~/.q3e`). It also skips the logo which is not an expected behavior (because the in-game console is triggered at every launch to set of the command arg).