Better command args for ioquake3 and quake3e to be at screen resolution (#205)
* Remove "r_mode 2" for ioquake3 and quake3e

The command argument `r_mode -2` for both ioquake3 and quake3e is redundant given that it's the default mode when the config file is created (in `~/.q3e`). It also skips the logo which is not an expected behavior (because the in-game console is triggered at every launch to set of the command arg).

* Update packages.json

1/ It's better to use the command (for example) `+set r_mode -2` instead of `+r_mode -2`, the first one will set the arg before launching the game, it's not necessary to restart the game and it doesn't skip the logo (unless with the specific command)

2/ ioquake3 and quake3e don't have the same default modes when it comes to mods, which Team Arena is one of them:

ioquake3 with `+set fs_game missionpack` by default:
`r_mode 3`
`r_modefullscreen` not present

quake3e with `+set fs_game missionpack` by default:
`r_mode 3`
`r_modefullscreen -2`

To ensure to be always at screen resolution by default with the main game and its mods, I think the best way is to use the command args `+set r_mode -2 +set r_modeFullscreen -2`. If the user wants something different, it remains possible to do it in the launch option of the game in Steam.